# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IUserAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"dni": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true
					},
					"dniType": {
						"type": "string",
						"nullable": true
					},
					"phoneNumber": {
						"type": "string",
						"nullable": true
					},
					"dob": {
						"type": "string",
						"nullable": true
					},
					"password": {
						"type": "string"
					},
					"rolId": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "boolean"
					},
					"location": {
						"type": "string",
						"nullable": true
					},
					"tokenPush": {
						"type": "string",
						"nullable": true
					},
					"receiveNotification": {
						"type": "boolean"
					},
					"gender": {
						"type": "string",
						"nullable": true
					},
					"state": {
						"type": "string",
						"nullable": true
					},
					"city": {
						"type": "string",
						"nullable": true
					},
					"zone": {
						"type": "string",
						"nullable": true
					},
					"uid": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"rolId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "apiKey",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "api-corpoindustri",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/views/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operación.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar una historia por ID.",
				"tags": [
					"View"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de la historia a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users/update/{userId}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdatenAttributes"
							}
						}
					}
				}
			}
		},
...
		"/users/updateTokenPush": {
			"put": {
				"operationId": "UpdateTokenPush",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Actualizar el token push de un usuario.",
				"tags": [
					"User"
				],
				"security": [
					{
						"bearerAuth": [
							"optional"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tokenPush": {
										"type": "string"
									}
								},
								"required": [
									"tokenPush"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar un usuario por ID.",
				"tags": [
					"User"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de usuario a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/rols/deleted/{key}": {
			"delete": {
				"operationId": "SoftDeleteRecord",
				"responses": {
					"200": {
						"description": "- Resultado de la operacion",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"summary": "Eliminar un usuario por ID.",
				"tags": [
					"Rol"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "- ID de usuario a eliminar.",
						"in": "path",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"user": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"user",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Registra un nuevo usuario.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreationAttributes"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Usuario autenticado y token de autenticación o mensaje de error.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"token": {
											"type": "string"
										},
										"user": {
											"$ref": "#/components/schemas/IUserAttributes"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Inicia sesión con credenciales de usuario.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Credenciales de inicio de sesión (email y contraseña).",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"recaptchaToken": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"recaptchaToken",
									"password",
									"email"
								],
								"type": "object",
								"description": "- Credenciales de inicio de sesión (email y contraseña)."
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "Me",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Obtiene los detalles del usuario autenticado.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/auth/password": {
			"patch": {
				"operationId": "PasswordUpdate",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"user": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"user",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Obtiene los detalles del usuario autenticado.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": [
							"optional"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Solicitud HTTP con información de autenticación.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"oldPasword": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"passwordConfirmation": {
										"type": "string"
									}
								},
								"required": [
									"oldPasword",
									"password",
									"passwordConfirmation"
								],
								"type": "object",
								"description": "- Solicitud HTTP con información de autenticación."
							}
						}
					}
				}
			}
		},
		"/auth/update/{userId}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": [
							"optional"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthUserUpdatenAttributes"
							}
						}
					}
				}
			}
		},
		"/auth/updateTokenPush": {
			"patch": {
				"operationId": "UpdateTokenPush",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"item": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"item",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Actualizar el token push.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": [
							"optional"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- tokenPush en formato texto a actualizar.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tokenPush": {
										"type": "string"
									}
								},
								"required": [
									"tokenPush"
								],
								"type": "object",
								"description": "- tokenPush en formato texto a actualizar."
							}
						}
					}
				}
			}
		},
		"/auth/passwordReset": {
			"post": {
				"operationId": "PasswordReset",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Obtiene los detalles del usuario autenticado.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Solicitud HTTP con información de autenticación.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object",
								"description": "- Solicitud HTTP con información de autenticación."
							}
						}
					}
				}
			}
		},
		"/auth/passwordRecovery": {
			"post": {
				"operationId": "PasswordRecovery",
				"responses": {
					"201": {
						"description": "User Found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"token": {
											"type": "string"
										},
										"user": {
											"allOf": [
												{
													"$ref": "#/components/schemas/IUserAttributes"
												}
											],
											"nullable": true
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"user",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Cambiar la contraseña, por codigo (al recuperar) o por contraseña anterior.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Solicitud HTTP con información de autenticación.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPasswordRecovery",
								"description": "- Solicitud HTTP con información de autenticación."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}